fix: rename filenames to camelCase
