+++ app/assets/javascripts/preferences/panes/security.tsx
@@ -2,7 +2,7 @@ import { observer } from 'mobx-react-lite';
 import { FunctionComponent } from 'preact';
 import { PreferencesPane } from '../components';
 import { Preferences } from '../models';
-import { TwoFactorAuthComponent } from './two-factor-auth';
+import { TwoFactorAuthComponent } from './twoFactorAuth';
 
 export const Security: FunctionComponent<{ prefs: Preferences }> = observer(
   ({ prefs }) => (
